+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/arcon/compare/v0.2.1...HEAD)
+[Unreleased](https://github.com/jshwi/arcon/compare/v0.3.0...HEAD)
+------------------------------------------------------------------------
+
+[0.3.0](https://github.com/jshwi/arcon/releases/tag/v0.3.0) - 2023-09-17
 ------------------------------------------------------------------------
 ### Added
 - add param to override pyproject.toml config